+++ src/main/kotlin/travesium/userservice/UserServiceApplication.kt
@@ -1,12 +1,13 @@
 package travesium.userservice
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.autoconfigure.kafka.KafkaAutoConfiguration
 import org.springframework.boot.runApplication
 import org.springframework.context.annotation.Import
 import traversium.commonmultitenancy.FlywayTenantMigration
 import traversium.commonmultitenancy.MultiTenantAutoConfiguration
 
-@SpringBootApplication
+@SpringBootApplication(exclude = [KafkaAutoConfiguration::class])
 @Import(MultiTenantAutoConfiguration::class, FlywayTenantMigration::class)
 class UserServiceApplication