+++ src/main/kotlin/travesium/userservice/kafka/KafkaConfig.kt
@@ -9,6 +9,7 @@ import org.apache.kafka.common.serialization.StringSerializer
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.boot.context.properties.EnableConfigurationProperties
+import org.springframework.cloud.context.config.annotation.RefreshScope
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.core.env.Environment
@@ -26,6 +27,7 @@ import java.util.concurrent.TimeUnit
 @EnableKafka
 @EnableConfigurationProperties(KafkaProperties::class)
 @ConditionalOnProperty(prefix = "spring.kafka", name = ["bootstrap-servers"])
+@RefreshScope
 class KafkaConfig {
 
     @Bean
@@ -47,7 +49,7 @@ class KafkaConfig {
     }
 
     @Bean
-    fun kafkaTemplate(
+    fun customKafkaTemplate(
         producerFactory: ProducerFactory<String, Any>,
         @Value("\${${KAFKA_PROPERTY_PREFIX}health.indicator.timeout.ms:1000}")
         kafkaTimeoutForHealthEndpoint: Long