+++ src/main/kotlin/travesium/userservice/kafka/KafkaProperties.kt
@@ -1,23 +1,23 @@
 package travesium.userservice.kafka
 
+import jakarta.annotation.PostConstruct
 import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.boot.context.properties.bind.ConstructorBinding
 
 /**
  * @author Maja Razinger
  */
 @ConfigurationProperties(prefix = "spring.kafka")
-class KafkaProperties @ConstructorBinding constructor(
-    val bootstrapServers: String,
-    val reportingTopic: String?,
-    val notificationTopic: String?,
-    val auditTopic: String?,
-    partition: Int? = null,
-    partitioningStrategy: PartitioningStrategy = PartitioningStrategy.PER_MESSAGE_KEY,
-    val clientConfirmationTimeout: Long = 10L
-) {
+class KafkaProperties {
+    lateinit var bootstrapServers: String
+    var reportingTopic: String? = null
+    var notificationTopic: String? = null
+    var auditTopic: String? = null
+    var partition: Int? = null
+    var partitioningStrategy: PartitioningStrategy = PartitioningStrategy.PER_MESSAGE_KEY
+    var clientConfirmationTimeout: Long = 10L
 
-    init {
+    @PostConstruct
+    fun validate() {
         if (partitioningStrategy === PartitioningStrategy.FIXED && partition == null) {
             throw IllegalArgumentException("Partition number must be defined in case of fixed partitioning strategy.")
         }