+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add circuit breaker and retry logic for external service calls
 - ELK stack integration
 - Add examples to swagger documentation
+- Support for config server
 
 ## [1.1.0] - 2025-12-03
 ### Added